pyproject.toml
[project]
name = "epub-optimizer-xteink"
version = "0.1.0"
description = "EPUB optimizer for Xteink X4 e-reader (480x800 e-paper display)"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
keywords = ["epub", "ebook", "e-reader", "xteink", "e-ink", "e-paper"]
dependencies = [
    "freetype-py>=2.5.1",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "isort>=6.0.1",
    "mypy>=1.16.0",
    "pyclean>=3.1.0",
    "pytest>=8.4.0",
    "pytest-aio>=1.9.0",
    "pytest-cov>=6.1.1",
    "pytest-mock>=3.14.1",
]

[tool.black]
line-length = 120
target-version = ['py312']
skip-string-normalization = true
include = '\.pyi?$'
exclude = 'build\/|buck-out\/|dist\/|_build\/|.git\/|.hg\/|.mypy_cache\/|.tox\/|.venv\/|migrations\/'

[project.scripts]
epub-to-xtc = "converter:main"

[project.urls]
Homepage = "https://github.com/bigbag/xteink-epub-optimizer"
Repository = "https://github.com/bigbag/xteink-epub-optimizer"
Issues = "https://github.com/bigbag/xteink-epub-optimizer/issues"
